rename context, close panel on step change
