+++ airbyte-webapp/src/pages/OnboardingPage/components/DestinationStep.tsx
@@ -9,7 +9,7 @@ import { useDestinationDefinitionList } from "services/connector/DestinationDefi
 import { useGetDestinationDefinitionSpecificationAsync } from "services/connector/DestinationDefinitionSpecificationService";
 import { createFormErrorMessage } from "utils/errorStatusMessage";
 import { ConnectorCard } from "views/Connector/ConnectorCard";
-import { useDocumentationPanelContext } from "views/Connector/ConnectorDocumentationLayout/ConnectorDocumentationContext";
+import { useDocumentationPanelContext } from "views/Connector/ConnectorDocumentationLayout/DocumentationPanelContext";
 
 import HighlightedText from "./HighlightedText";
 import TitlesBlock from "./TitlesBlock";
