+++ airbyte-webapp/src/views/Connector/ConnectorDocumentationLayout/ConnectorDocumentationWrapper.tsx
@@ -1,5 +1,5 @@
-import { DocumentationPanelProvider } from "./ConnectorDocumentationContext";
 import { ConnectorDocumentationLayout } from "./ConnectorDocumentationLayout";
+import { DocumentationPanelProvider } from "./DocumentationPanelContext";
 
 export const ConnectorDocumentationWrapper: React.FC = ({ children }) => {
   return (
